+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. Unreleased 
 - Roll command with button interaction: parses the message (or button interaction) ID for repeating last digits (e.g. 881101126133399**000** )
 - SQLite3 support
 - Triggerbyid command
+- Support for higher upload limits (server boost levels 2 and 3)
 
 ### Changed
 - Reminder response formatting